+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 
 
+## 0.3.2 (2024-11-06)
+
+## 0.3.1 (2024-10-28)
+
+## 0.3.0 (2024-10-25)
+
+
+### Features
+
+* add deploy contract action ([#17](https://github.com/yeagerai/genlayer-js/issues/17)) ([23d5bc2](https://github.com/yeagerai/genlayer-js/commit/23d5bc28fb58c73d64b1fd629185a0565d84cb91))
+
+## 0.2.0 (2024-10-17)
+
+
+### Features
+
+* implement genvm calldata ([#14](https://github.com/yeagerai/genlayer-js/issues/14)) ([d9a1abd](https://github.com/yeagerai/genlayer-js/commit/d9a1abdfb5eef13e5c77433db546953369087e04))
+
 ## 0.1.2 (2024-10-03)
 
 ## 0.1.1 (2024-10-03)